pyproject.toml
[tool.poetry]
name = "asktowebapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.4.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
dataclasses-json = "0.6.7"
deprecation = "2.1.0"
distro = "1.9.0"
dnspython = "2.6.1"
email-validator = "2.1.2"
exceptiongroup = "1.2.1"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
frozenlist = "1.4.1"
gotrue = "2.4.4"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
jinja2 = "3.1.4"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain = "0.2.3"
langchain-community = "0.2.4"
langchain-core = "0.2.5"
langchain-cohere = "0.1.8"
langchain-openai = "0.1.8"
langchain-pinecone = "0.1.1"
langchain-text-splitters = "0.2.1"
langchain-anthropic = "0.1.15"
langsmith = "0.1.77"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.21.3"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
openai = "1.32.0"
orjson = "3.10.5"
packaging = "23.2"
pinecone-client = "3.2.2"
postgrest = "0.16.8"
pydantic = "2.7.4"
pydantic-core = "2.18.4"
pygments = "2.18.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
realtime = "1.0.6"
regex = "2024.5.15"
requests = "2.32.3"
rich = "13.7.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.30"
starlette = "0.37.2"
storage3 = "0.7.6"
strenum = "0.4.15"
supabase = "2.5.0"
supafunc = "0.4.6"
tenacity = "8.3.0"
tiktoken = "0.7.0"
tqdm = "4.66.4"
typer = "0.12.3"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
ujson = "5.10.0"
urllib3 = "2.2.1"
uvicorn = "0.30.1"
uvloop = "0.19.0"
watchfiles = "0.22.0"
websockets = "12.0"
yarl = "1.9.4"
langchain-groq = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
